move ModalCloseButton to constants.ts
