--- packages/vlossom/src/components/vs-modal/stories/ModalCloseButton.ts
@@ -1,9 +1,0 @@
-import { defineComponent } from 'vue';
-import VsButton from '@/components/vs-button/VsButton.vue';
-
-export const ModalCloseButton = defineComponent({
-    components: { VsButton },
-    template: `
-        <vs-button>Close</vs-button>
-    `,
-});
