+++ packages/vlossom/src/components/vs-modal/stories/constants.ts
@@ -1,3 +1,12 @@
+import VsButton from '@/components/vs-button/VsButton.vue';
+
+export const ModalCloseButton = {
+    components: { VsButton },
+    template: `
+        <vs-button>Close</vs-button>
+    `,
+};
+
 export const containerStyle = {
     backgroundColor: 'rgba(0, 0, 0, 0.02)',
     border: '1px solid rgb(240, 240, 240)',
